# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Liferay ↔ Zoho CRM Client Extension API
  description: |
    RESTful API that bridges Liferay DXP and Zoho CRM.
    Exposes CRUD operations for Leads, Contacts, and Deals,
    plus bulk sync endpoints and a real-time webhook receiver.
  version: 1.0.0
  contact:
    name: Your Organisation
    email: dev@your-org.com
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-liferay-host/o/zoho-crm
    description: Production (via Liferay gateway)
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/leads:
    get:
      tags: [Leads]
      summary: List leads (paginated)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Fields'
        - $ref: '#/components/parameters/Criteria'
        - $ref: '#/components/parameters/NoCache'
      responses:
        '200':
          description: Paginated list of leads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Leads]
      summary: Create or upsert lead(s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Lead'
                - type: array
                  items:
                    $ref: '#/components/schemas/Lead'
      responses:
        '201':
          description: Lead(s) created
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Fields:
      name: fields
      in: query
      schema:
        type: string
        example: "firstName,lastName,email"
    Criteria:
      name: criteria
      in: query
      schema:
        type: string
        example: "(Email:equals:test@example.com)"
    NoCache:
      name: nocache
      in: query
      schema:
        type: boolean
  schemas:
    Lead:
      type: object
      required: [lastName]
      properties:
        firstName:    { type: string }
        lastName:     { type: string }
        email:        { type: string, format: email }
        phone:        { type: string }
        company:      { type: string }
        title:        { type: string }
        leadSource:   { type: string }
        industry:     { type: string }
        description:  { type: string }
        city:         { type: string }
        state:        { type: string }
        country:      { type: string }
...
    LeadListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Lead'
        pagination:
          type: object
          properties:
            page:         { type: integer }
            per_page:     { type: integer }
            count:        { type: integer }
            more_records: { type: boolean }
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  type: object